separate the interface logic from the core logic into different files this way the core can be required separately and new interfaces would be easier to implement
